requirements.txt
fastapi
uvicorn
pdfplumber
beautifulsoup4
pytesseract
pillow
requests
sentence-transformers
azure-search-documents
matplotlib

msrest
azure-cognitiveservices-vision-computervision


torch
torchvision

